requirements.txt
Brotli==1.0.9
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dash==2.6.1
dash-auth==1.3.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Compress==1.12
Flask-SeaSurf==1.1.1
fpdf==1.7.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaleido==0.0.3.post1
MarkupSafe==2.1.1
numpy==1.23.3
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.2.0
plotly==5.10.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
retrying==1.3.3
six==1.16.0
tenacity==8.0.1
ua-parser==0.16.1
urllib3==1.26.12
Werkzeug==2.2.2
